// src/navigation/types.ts
export type MainStackParamList = {
  Pokedex: undefined;
  PokemonDetail: {
    pokemonId: number;
    name: string;
  };
  Hunt: undefined;
  ARCamera: undefined;
};

export type TabParamList = {
  PokedexTab: undefined;
  HuntTab: undefined;
  ARCameraTab: undefined;
};

export type AuthStackParamList = {
  SignIn: undefined;
  SignUp: undefined;
};